% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axial_helper_functions.R
\name{axpermute}
\alias{axpermute}
\title{Estimate the axial dispersal distance of a kernel with confidence intervals}
\usage{
axpermute(vals, nreps = 1000, nsamp = "std", composite = 2, output = "confs")
}
\arguments{
\item{vals}{numeric. Vector of distances between close kin OR object of class KinPairData.}

\item{nreps}{numeric. Number of permutations to run for confidence intervals (default 1000)}

\item{nsamp}{numeric. Number of kin pairs to subsample for each permutation.
Either "std" or an integer. If "std" will be computed as equal to the sample size. (default "std")}

\item{composite}{numeric. Number of 'draws' going into each underlying distribution (default 2). Passed to axials function.}

\item{output}{character. Denotes what kind of output to return.
If 'confs', a vector of 95\% confidence intervals. if 'vect', a vector of all permuted axial value results}
}
\value{
If ouput = 'confs', returns a vector of 95\% confidence intervals.
if output = 'vect', returns a vector of all permuted axial value results
}
\description{
Estimate the axial dispersal distance of a kernel with confidence intervals
}
\examples{
po_dists <- rexp(100, 1 / 50)
axpermute(po_dists, composite = 1)
}
