% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axial_helper_functions.R
\name{axpermute_subtract}
\alias{axpermute_subtract}
\title{Decompose distributions with confidence intervals}
\usage{
axpermute_subtract(
  bigvals,
  smallvals,
  nreps = 1000,
  num = 50,
  composite = 2,
  output = "confs"
)
}
\arguments{
\item{output}{}
}
\value{

}
\description{
Find the difference between two different axial distributions with confidence intervals.
}
