% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\name{breeding_cycle}
\alias{breeding_cycle}
\title{kindisperse - access breeding cycle at sampling of DispersalModel object.}
\usage{
breeding_cycle(x)
}
\arguments{
\item{x}{object of class \code{DispersalModel} or \code{KinPairData}}
}
\value{
\code{non-negative integer(s)} Breeding cycle numbers of modelled dispersed kin. Represents the number of complete
breeding cycles each indivdiual has undergone before the sampling point, wher ethe time between birth and first
reproduction is coded as \code{0}, that between first and second reproduction \code{1}, etc.
}
\description{
kindisperse - access breeding cycle at sampling of DispersalModel object.
}
