% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\name{breeding_stage}
\alias{breeding_stage}
\alias{breeding_stage,DispersalModel-method}
\title{Access life stage at which breeding occurs of \code{\link{DispersalModel}} object}
\usage{
breeding_stage(x)

\S4method{breeding_stage}{DispersalModel}(x)
}
\arguments{
\item{x}{object of class \code{DispersalModel}}

\item{DispersalModel}{object of class \code{DispersalModel}}
}
\value{
\code{character} life stage at which breeding occurs for modeled dispersed kin.
}
\description{
Access life stage at which breeding occurs of \code{\link{DispersalModel}} object
}
\section{Methods (by class)}{
\itemize{
\item \code{DispersalModel}: 
}}

