% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{breedsigma<-}
\alias{breedsigma<-}
\title{Title}
\usage{
breedsigma(x) <- value
}
\arguments{
\item{x}{object with relevant method}

\item{value}{new value to assign}
}
\value{
returns a modified object of the relevant class
}
\description{
Title
}
