% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{breedsigma}
\alias{breedsigma}
\title{Access breedsigma (KinPairSimulation method)}
\usage{
\S4method{breedsigma}{KinPairSimulation}(x)
}
\arguments{
\item{KinPairSimulation}{}
}
\value{

}
\description{
Access breedsigma (KinPairSimulation method)
}
