% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairData.R
\name{category}
\alias{category}
\alias{category<-,KinPairData-method}
\title{Title}
\usage{
category(x)

\S4method{category}{KinPairData}(x)

\S4method{category}{KinPairData}(x) <- value
}
\arguments{
\item{x}{}

\item{KinPairData}{}
}
\value{





}
\description{
Title

Access or assign KinPairData kin category
}
