% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_kinpairs.R
\name{check_valid_kinship}
\alias{check_valid_kinship}
\title{Check if vector of kinship categories contains all valid entries}
\usage{
check_valid_kinship(vect)
}
\arguments{
\item{vect}{vector of kinship categories}
}
\value{
TRUE if valid. Error otherwise.
}
\description{
Check if vector of kinship categories contains all valid entries
}
