% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_model.R
\name{dispersal_model}
\alias{dispersal_model}
\title{Build species-specific dispersal model with custom lifestage, phase & sampling parameters}
\usage{
dispersal_model(..., .FS = 0, .HS = .FS, .lifestage = 0, .cycle = 0)
}
\arguments{
\item{...}{name, value (numeric) pairs pairing custom lifestages with their corresponding axial dispersal values. MUST
be in chronological order across the entire breeding cycle.}

\item{.FS}{(character) - breeding cycle stage at which first substantial FS-phased dispersal occurs. Must correspond to a
previously described cycle stage name. Typically reflects the first dispersal of female gametes from the mother at (variously)
egg-laying, birth, weaning stages (species-dependent). Use care in adapting to situations where multiple breeding and/or dispersal
routes commonly lead to the FS phase}

\item{.HS}{(character) - breeding cycle stage at which first substantial HS-phased dispersal occurs. Must correspond to a
previously described cycle stage name. Typically reflects the movement of male gametes at e.g. the breeding stage (use care in adapting
to situations where multiple dispersal routes commonly lead to the HS phase)}

\item{.lifestage}{(character) - stage in the breeding cycle at which samples are to be collected for kin identification. Must correspond
to a previously described cycle stage name. (so collection of eggs corresponds to an egg-laying stage, as juveniles to a juvenile stage, etc.)}

\item{.cycle}{(non-negative integer or vector of two non-negative integers) breeding cycle numbers of dispersed kin to be modeled. Represents
the number of complete breeding cycles each simulated individual has undergone before the sampling point, where the time between
birth and first reproduction is coded as '0', that between first and second reproduction '1', etc. (default 0). Only use in spp.
where there is likely to be a reasonable equivalence between breeding stages across a lifespan. As the rest of the model is compatible
with a variety of cycle points, this parameter will often be overridden by the 'cycle' parameter in the \code{simulate_kindist_custom} function.}
}
\value{
Returns an object of class \code{DispersalModel} containing custom lifestages and dispersal, phase & sampling parameters that
can be passed to simulation functions.
}
\description{
Build species-specific dispersal model with custom lifestage, phase & sampling parameters
}
\examples{
antechinus_model <- dispersal_model(pouch = 25, nest = 25, free_living = 250, breeding = 40,
gestation = 25, .FS = "nest", .HS = "breeding", .lifestage = "nest")
antechinus_model
}
