% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\name{dispersal_vector}
\alias{dispersal_vector}
\title{kindisperse - access dispersal vector of DispersalModel object.}
\usage{
dispersal_vector(x)
}
\arguments{
\item{x}{object of class \code{DispersalModel}}
}
\value{
\code{numeric vector} named vector of custom lifestages & associated dispersal sigmas.
}
\description{
kindisperse - access dispersal vector of DispersalModel object.
}
