% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\docType{class}
\name{DispersalModel-class}
\alias{DispersalModel-class}
\alias{DispersalModel}
\alias{dispersal_vector,DispersalModel-method}
\alias{stages,DispersalModel-method}
\alias{fs,DispersalModel-method}
\alias{hs,DispersalModel-method}
\alias{sampling_stage,DispersalModel-method}
\alias{breeding_cycle,DispersalModel-method}
\alias{show,DispersalModel-method}
\alias{initialize,DispersalModel-method}
\title{DispersalModel Class}
\usage{
\S4method{dispersal_vector}{DispersalModel}(x)

\S4method{stages}{DispersalModel}(x)

\S4method{fs}{DispersalModel}(x)

\S4method{hs}{DispersalModel}(x)

\S4method{sampling_stage}{DispersalModel}(x)

\S4method{breeding_cycle}{DispersalModel}(x)

\S4method{show}{DispersalModel}(object)

\S4method{initialize}{DispersalModel}(
  .Object,
  stages = NULL,
  dispersal_vector = NULL,
  fs = NULL,
  hs = NULL,
  sampling_stage = NULL,
  cycle = NULL
)
}
\arguments{
\item{x}{object of class \code{DispersalModel}}

\item{object}{object of class \code{DispersalModel}}

\item{.Object}{object to be constructed into DispersalModel class}

\item{stages}{character.  Ordered vector of all dispersal stages across the breeding cycle of the modeled species}

\item{dispersal_vector}{numeric. Named vector of custom breeding cycle stages and their corresponding axial dispersal values}

\item{fs}{character.  breeding cycle stage at which first substantial FS-phased dispersal occurs}

\item{hs}{character. breeding cycle stage at which first substantial HS-phased dispersal occurs}

\item{sampling_stage}{character. stage in the breeding cycle at which samples are to be collected for kin identification.}

\item{cycle}{non-negative integer. Breeding cycle numbers of dispersed kin to be modeled.  Represents
the number of complete breeding cycles each simulated individual has undergone before the sampling point, where the time between
birth and first reproduction is coded as '0', that between first and second reproduction '1', etc. (default 0)}

\item{DispersalModel}{an object of class DispersalModel}
}
\value{
returns object of class \code{DispersalModel}

\code{numeric vector} named vector of custom lifestages & associated dispersal sigmas.

\code{character} ordered vector of custom lifestages contained in the object

\code{character} FS phase split

\code{character} HS phase split

\code{character} sampling stage

\code{non-negative integer(s)} Breeding cycle numbers of modelled dispersed kin. Represents the number of complete
breeding cycles each indivdiual has undergone before the sampling point, wher ethe time between birth and first
reproduction is coded as \code{0}, that between first and second reproduction \code{1}, etc.

No return value. Called for side effects

returns an object of class \code{DispersalModel}
}
\description{
DispersalModel Class
}
\section{Methods (by generic)}{
\itemize{
\item \code{dispersal_vector}: access dispersal vector

\item \code{stages}: access breeding cycle stages

\item \code{fs}: access FS phase split

\item \code{hs}: access HS phase split

\item \code{sampling_stage}: access sampling stage

\item \code{breeding_cycle}: access breeding cycle at sampling

\item \code{show}: print method

\item \code{initialize}: initialization method
}}

\section{Slots}{

\describe{
\item{\code{dispersal_vector}}{numeric. Named vector of custom breeding cycle stages and their corresponding axial dispersal values}

\item{\code{stages}}{character.  Ordered vector of all dispersal stages across the breeding cycle of the modelled species}

\item{\code{fs}}{character.  breeding cycle stage at which first substantial FS-phased dispersal occurs}

\item{\code{hs}}{character. breeding cycle stage at which first substantial HS-phased dispersal occurs}

\item{\code{sampling_stage}}{character. stage in the breeding cycle at which samples are to be collected for kin identification.}

\item{\code{cycle}}{non-negative integer. Breeding cycle numbers of dispersed kin to be modeled.  Represents
the number of complete breeding cycles each simulated individual has undergone before the sampling point, where the time between
birth and first reproduction is coded as '0', that between first and second reproduction '1', etc. (default 0)}
}}

