% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ports.R
\name{display_appdata}
\alias{display_appdata}
\title{Show printout of named items stored in appdata.}
\usage{
display_appdata()
}
\value{
No return value, called for side effects
}
\description{
Show printout of named items stored in appdata.
}
\examples{
mount_appdata(KinPairData(), "my_kindata")
mount_appdata(simulate_kindist_simple(nsims = 10), "my_simdata")

display_appdata()
}
