% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{filter_methods}
\alias{filter_methods}
\alias{upper}
\alias{upper<-}
\alias{lower}
\alias{lower<-}
\alias{spacing}
\alias{spacing<-}
\alias{samplenum}
\alias{samplenum<-}
\alias{sampledims}
\alias{sampledims<-}
\alias{upper,KinPairSimulation-method}
\alias{lower,KinPairSimulation-method}
\alias{spacing,KinPairSimulation-method}
\alias{samplenum,KinPairSimulation-method}
\alias{sampledims,KinPairSimulation-method}
\alias{upper<-,KinPairSimulation-method}
\alias{lower<-,KinPairSimulation-method}
\alias{spacing<-,KinPairSimulation-method}
\alias{samplenum<-,KinPairSimulation-method}
\alias{sampledims<-,KinPairSimulation-method}
\title{Access or modify the \code{filter} parameters of \code{\link{KinPairSimulation}} objects}
\usage{
upper(x)

upper(x) <- value

lower(x)

lower(x) <- value

spacing(x)

spacing(x) <- value

samplenum(x)

samplenum(x) <- value

sampledims(x)

sampledims(x) <- value

\S4method{upper}{KinPairSimulation}(x)

\S4method{lower}{KinPairSimulation}(x)

\S4method{spacing}{KinPairSimulation}(x)

\S4method{samplenum}{KinPairSimulation}(x)

\S4method{sampledims}{KinPairSimulation}(x)

\S4method{upper}{KinPairSimulation}(x) <- value

\S4method{lower}{KinPairSimulation}(x) <- value

\S4method{spacing}{KinPairSimulation}(x) <- value

\S4method{samplenum}{KinPairSimulation}(x) <- value

\S4method{sampledims}{KinPairSimulation}(x) <- value
}
\arguments{
\item{x}{object of class KinPairSimulation}

\item{value}{value for parameter to be adjusted to}

\item{KinPairSimulation}{object of class KinPairSimulation}
}
\value{
either the accessed \code{numeric} filter parameter or a filtered \code{\link{KinPairSimulation}} object
}
\description{
These generics & methods work as an interface between \code{KinPairSimulation} objects and the
\code{\link{sample_kindist}} function. They either retrieve the value of pre-existing filter steps that have
been applied to the object (e.g. \code{upper(x)}) or assign such a filtering parameter to the \code{KinPairSimulation}
object (e.g. \code{sampledims(x) <- value}). In this case, the method passes the \code{KinPairSimulation} object to the
\code{sample_kindist()} function for subsampling or filtering, then updates the sampling parameter before returning the
modified object. Note that while the \code{sample_kindist} function can take \code{\link{KinPairData}} objects, the methods
described here are only applicable to objects of class \code{KinPairSimulation}.
}
\section{Functions}{
\itemize{
\item \code{upper,KinPairSimulation-method}: 

\item \code{lower,KinPairSimulation-method}: 

\item \code{spacing,KinPairSimulation-method}: 

\item \code{samplenum,KinPairSimulation-method}: 

\item \code{sampledims,KinPairSimulation-method}: 

\item \code{upper<-,KinPairSimulation-method}: 

\item \code{lower<-,KinPairSimulation-method}: 

\item \code{spacing<-,KinPairSimulation-method}: 

\item \code{samplenum<-,KinPairSimulation-method}: 

\item \code{sampledims<-,KinPairSimulation-method}: 
}}

\seealso{
Other kpsmethods: 
\code{\link{access_sigmas}},
\code{\link{kernelshape}()},
\code{\link{kerneltype}()},
\code{\link{simtype}()}
}
\concept{kpsmethods}
