% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{filtertype}
\alias{filtertype}
\title{Access filter type of KinPairSimulation object}
\usage{
filtertype(x)
}
\arguments{
\item{x}{object of class \code{KinPairSimulation}}
}
\value{

}
\description{
Access filter type of KinPairSimulation object
}
