% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\name{fs}
\alias{fs}
\title{kindisperse - access FS phase split point of DispersalModel object.}
\usage{
fs(x)
}
\arguments{
\item{x}{object of class \code{DispersalModel}}
}
\value{
\code{character} FS phase split
}
\description{
kindisperse - access FS phase split point of DispersalModel object.
}
