% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{get_dispersal_model}
\alias{get_dispersal_model}
\alias{get_dispersal_model,KinPairSimulation-method}
\title{Access dispersal model of \code{\link{KinPairSimulation}} object}
\usage{
get_dispersal_model(x)

\S4method{get_dispersal_model}{KinPairSimulation}(x)
}
\arguments{
\item{x}{object of class \code{KinPairSimulation}}

\item{KinPairSimulation}{object of class \code{KinPairSimulation}}
}
\value{
returns an object of class \code{\link{DispersalModel}}
}
\description{
Access dispersal model of \code{\link{KinPairSimulation}} object
}
\section{Methods (by class)}{
\itemize{
\item \code{KinPairSimulation}: 
}}

