% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{gravsigma}
\alias{gravsigma}
\title{Access gravsigma (KinPairSimulation method)}
\usage{
\S4method{gravsigma}{KinPairSimulation}(x)
}
\arguments{
\item{KinPairSimulation}{}
}
\value{

}
\description{
Access gravsigma (KinPairSimulation method)
}
