% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\name{hs}
\alias{hs}
\title{kindisperse - access HS phase split point of DispersalModel object.}
\usage{
hs(x)
}
\arguments{
\item{x}{object of class \code{DispersalModel}}
}
\value{
\code{character} HS phase split
}
\description{
kindisperse - access HS phase split point of DispersalModel object.
}
