% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\name{is.DispersalModel}
\alias{is.DispersalModel}
\title{Check if object is of class \code{DispersalModel}}
\usage{
is.DispersalModel(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
returns TRUE if of class \code{DispersalModel}, FALSE if not
}
\description{
Check if object is of class \code{DispersalModel}
}
