% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairData.R
\name{is.KinPairData}
\alias{is.KinPairData}
\title{Check if object is of class KinPairData}
\usage{
is.KinPairData(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
Returns TRUE if of class \code{KinPairData}, FALSE if not.
}
\description{
Check if object is of class KinPairData
}
