% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{is.KinPairSimulation}
\alias{is.KinPairSimulation}
\title{Check if object is of class KinPairSimulation}
\usage{
is.KinPairSimulation(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
Returns TRUE if of class KinPairSimulation, FALSE if not
}
\description{
Check if object is of class KinPairSimulation
}
