% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{juvsigma<-}
\alias{juvsigma<-}
\title{Title}
\usage{
juvsigma(x) <- value
}
\arguments{
\item{x}{object of class \code{KinPairSimulation}}

\item{value}{new value to assign}
}
\value{

}
\description{
Title
}
