% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{juvsigma}
\alias{juvsigma}
\title{Access juvsigma (KinPairSimulation method)}
\usage{
\S4method{juvsigma}{KinPairSimulation}(x)
}
\arguments{
\item{KinPairSimulation}{}
}
\value{

}
\description{
Access juvsigma (KinPairSimulation method)
}
