% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{kernelshape}
\alias{kernelshape}
\title{kindisperse - access kerneltype of KinPairSimulation object}
\usage{
kernelshape(x)
}
\arguments{
\item{x}{object of class \code{KinPairSimulation}}
}
\value{

}
\description{
kindisperse - access kerneltype of KinPairSimulation object
}
