% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{kerneltype}
\alias{kerneltype}
\title{Access kerneltype (KinPairSimulation method)}
\usage{
\S4method{kerneltype}{KinPairSimulation}(x)
}
\value{

}
\description{
Access kerneltype (KinPairSimulation method)
}
