% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{kinpair_to_csv}
\alias{kinpair_to_csv}
\title{Write KinPairData object in .csv format (strips extra data)}
\usage{
kinpair_to_csv(x, file, ...)
}
\arguments{
\item{x}{Oject of class KinPairData or KinPairSimulation}

\item{file}{The file path to write to}

\item{...}{Additional arguments to pass to write_csv}
}
\value{
Invisibly returns the initial object
}
\description{
Write KinPairData object in .csv format (strips extra data)
}
