% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairData.R
\name{to_tibble,KinPairData-method}
\alias{to_tibble,KinPairData-method}
\alias{distances,KinPairData-method}
\alias{kinship,KinPairData-method}
\alias{kinship<-,KinPairData-method}
\alias{lifestage,KinPairData-method}
\alias{lifestage<-,KinPairData-method}
\alias{show,KinPairData-method}
\alias{initialize,KinPairData-method}
\title{Title}
\usage{
\S4method{to_tibble}{KinPairData}(x)

\S4method{distances}{KinPairData}(x)

\S4method{kinship}{KinPairData}(x)

\S4method{kinship}{KinPairData}(x) <- value

\S4method{lifestage}{KinPairData}(x)

\S4method{lifestage}{KinPairData}(x) <- value

\S4method{show}{KinPairData}(object)

\S4method{initialize}{KinPairData}(.Object, data = NULL, kinship = NULL, lifestage = NULL, ...)
}
\arguments{
\item{KinPairData}{}
}
\value{















}
\description{
Title
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_tibble}: extract to tibble

\item \code{distances}: access distances

\item \code{kinship}: access kin category

\item \code{kinship<-}: assign kin category

\item \code{lifestage}: access lifestage

\item \code{lifestage<-}: assign lifestage

\item \code{show}: standard print method

\item \code{initialize}: initialize method
}}

