% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{KinPairSimulation_simple}
\alias{KinPairSimulation_simple}
\title{Constructor for KinPairSimulation Class (simple)}
\usage{
KinPairSimulation_simple(
  data = NULL,
  category = NULL,
  kerneltype = NULL,
  sigma = NULL,
  dims = NULL,
  lifestage = NULL,
  call = NULL
)
}
\arguments{
\item{data}{tibble of pairwise kin classes & distances. Ideally contains fields id1 & id2 (chr) an distance (dbl) optionally includes coords (x1, y1, x2, y2), lifestage (ls1 & ls2), category (chr) and sims (dbl)}

\item{category}{character. Code for kinship category of simulation. one of PO, FS, HS, AV, GG, HAV, GGG, 1C, 1C1, 2C, GAV, HGAV, H1C or H2C}

\item{kerneltype}{character. Statistical model for simulated dispersal kernel. Currently either "Gaussian" or "Laplace".}

\item{sigma}{numeric. Axial sigma of dispersal kernel (axial standard deviation).}

\item{dims}{numeric. Length of side of simulated area square.}

\item{lifestage}{character. Simulated lifestage of sampling. Either "larva" (sampled at hatching) or "oviposition" (sampled as an adult during oviposition - essentially one lifespan later that 'larva')}

\item{call}{call object. Use to pass the system call that led to the generation of this class. (via sys.call)}
}
\value{
Returns a KinPairSimulation Class object with simtype set to 'simple' and relevant fields included.
}
\description{
Constructor for KinPairSimulation Class (simple)
}
\examples{
kindata <- tibble::tibble(id1 = c("a", "b", "c"), id2 = c("x", "y", "z"),
    distance = c(50, 45, 65), category = c("1C", "1C", "1C"))
KinPairSimulation_simple(kindata, category = "1C", kerneltype = "Gaussian",
    sigma = 38, lifestage = "larva")
}
