% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{lower}
\alias{lower}
\alias{lower<-,KinPairSimulation-method}
\title{Access or filter by lower distance (KinPairSimulation method)}
\usage{
\S4method{lower}{KinPairSimulation}(x)

\S4method{lower}{KinPairSimulation}(x) <- value
}
\arguments{
\item{KinPairSimulation}{}
}
\value{



}
\description{
Access or filter by lower distance (KinPairSimulation method)
}
