% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{make_KinPairSimulation}
\alias{make_KinPairSimulation}
\title{Constructor for KinPairSimulation Class}
\usage{
make_KinPairSimulation(
  category,
  simtype,
  kerneltype,
  sigma,
  juvsigma,
  breedsigma,
  gravsigma,
  ovisigma,
  dims,
  lifestage,
  call,
  tab,
  filtertype,
  upper,
  lower,
  spacing,
  samplenum,
  sampledims
)
}
\arguments{
\item{sampledims}{}
}
\value{

}
\description{
Constructor for KinPairSimulation Class
}
