% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{ovisigma}
\alias{ovisigma}
\title{Access ovisigma (KinPairSimulation method)}
\usage{
\S4method{ovisigma}{KinPairSimulation}(x)
}
\arguments{
\item{KinPairSimulation}{}
}
\value{

}
\description{
Access ovisigma (KinPairSimulation method)
}
