% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{posigma}
\alias{posigma}
\alias{initsigma}
\alias{breedsigma}
\alias{gravsigma}
\alias{ovisigma}
\title{kindisperse - access sigmas of KinPairSimulation objects}
\usage{
posigma(x)

initsigma(x)

breedsigma(x)

gravsigma(x)

ovisigma(x)
}
\arguments{
\item{x}{object of class \code{KinPairSimulation}}
}
\value{
\code{numeric} posigma value of simple simulation

\code{numeric} initsigma value of composite simulation

\code{numeric} breedsigma value of composite simulation

\code{numeric} gravsigma value of composite simulation

\code{numeric} ovisigma value of composite simulation
}
\description{
kindisperse - access sigmas of KinPairSimulation objects
}
