% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_kindist.R
\name{rebase_dims}
\alias{rebase_dims}
\title{Change the dimensions of a KinPairSimulation Object and shift kinpairs so at least one individual is within the area}
\usage{
rebase_dims(kindist, dims)
}
\arguments{
\item{kindist}{\code{KinPairSimulation} - \code{KinPairSimulation} Class Object}

\item{dims}{New site dimensions - either single number (length of side of square) or length 2 vector (lengths of sides x and y of rectangle)}
}
\value{
returns a rebased object of class \code{KinPairSimulation} with adjusted simulation dimensions
}
\description{
Change the dimensions of a KinPairSimulation Object and shift kinpairs so at least one individual is within the area
}
\examples{
simobject <- simulate_kindist_simple()

rebase_dims(simobject, c(1, 100))
rebase_dims(simobject, 15)
}
