% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ports.R
\name{retrieveall_appdata}
\alias{retrieveall_appdata}
\title{Retrieve all KinPairData objects from appdata (as list)}
\usage{
retrieveall_appdata()
}
\value{

}
\description{
Retrieve all KinPairData objects from appdata (as list)
}
\examples{
mount_appdata(KinPairData(), "k1")
mount_appdata(KinPairSimulation(), "s1")
retrieveall_appdata()
}
