% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{sampledims}
\alias{sampledims}
\alias{sampledims<-,KinPairSimulation-method}
\title{Access or alter sampling dimensions (KinPairSimulation method)}
\usage{
\S4method{sampledims}{KinPairSimulation}(x)

\S4method{sampledims}{KinPairSimulation}(x) <- value
}
\arguments{
\item{KinPairSimulation}{}
}
\value{



}
\description{
Access or alter sampling dimensions (KinPairSimulation method)
}
