% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{samplenum}
\alias{samplenum}
\alias{samplenum<-,KinPairSimulation-method}
\title{Access or change sample number (KinPairSimulation method)}
\usage{
\S4method{samplenum}{KinPairSimulation}(x)

\S4method{samplenum}{KinPairSimulation}(x) <- value
}
\arguments{
\item{KinPairSimulation}{}
}
\value{



}
\description{
Access or change sample number (KinPairSimulation method)
}
