% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\name{sampling_stage}
\alias{sampling_stage}
\alias{sampling_stage,DispersalModel-method}
\title{Access sampling stage of \code{\link{DispersalModel}} object.}
\usage{
sampling_stage(x)

\S4method{sampling_stage}{DispersalModel}(x)
}
\arguments{
\item{x}{object of class \code{DispersalModel}}

\item{DispersalModel}{object of class \code{DispersalModel}}
}
\value{
\code{character} sampling stage
}
\description{
Access sampling stage of \code{\link{DispersalModel}} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{DispersalModel}: 
}}

