% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{sigma}
\alias{sigma}
\title{Access sigma (KinPairSimulation method)}
\usage{
\S4method{sigma}{KinPairSimulation}(x)
}
\arguments{
\item{KinPairSimulation}{}
}
\value{

}
\description{
Access sigma (KinPairSimulation method)
}
