% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{simdims<-}
\alias{simdims<-}
\title{Title}
\usage{
simdims(x) <- value
}
\arguments{
\item{x}{object of class \code{KinPairSimulation}}

\item{value}{new value to assign}
}
\value{
returns a modified object of the relevant class
}
\description{
Title
}
