% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{simtype<-}
\alias{simtype<-}
\title{Title}
\usage{
simtype(x) <- value
}
\arguments{
\item{x}{object of class \code{KinPairSimulation}}

\item{value}{new value to assign}
}
\value{
returns a modified object of relevant class
}
\description{
Title
}
