% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{simtype}
\alias{simtype}
\title{Access simtype (KinPairSimulation)}
\usage{
\S4method{simtype}{KinPairSimulation}(x)
}
\arguments{
\item{KinPairSimulation}{}
}
\value{

}
\description{
Access simtype (KinPairSimulation)
}
