% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_kindist_custom.R
\name{simulate_kindist_custom}
\alias{simulate_kindist_custom}
\title{Simulate kin dispersal distance pairs with custom species dispersal models.}
\usage{
simulate_kindist_custom(
  nsims = 100,
  model = dispersal_model(init = 100, breed = 50, grav = 50, ovi = 25, .FS = "ovi", .HS
    = "breed"),
  dims = 100,
  method = "Gaussian",
  kinship = "FS",
  lifestage = sampling_stage(model),
  cycle = 0,
  shape = 0.5
)
}
\arguments{
\item{nsims}{(integer) -   number of pairs to simulate}

\item{model}{(object of class \code{DispersalModel}) - species-specific model of dispersal containing lifestage & phase parameters}

\item{dims}{(numeric) -   length of sides of (square) simulated site area}

\item{method}{(character) - kernel shape to use: either 'Gaussian', 'Laplace' or 'vgamma' (variance-gamma)}

\item{kinship}{(character)- kin category to simulate: one of PO, FS, HS, AV, GG, HAV, GGG, 1C, 1C1, 2C, GAV, HGAV, H1C H1C1 or H2C}

\item{lifestage}{(character) lifestage at sample collection - must match a lifestage supplied in the 'model' parameter}

\item{cycle}{(numeric) - breeding cycle number(s) of dispersed kin to be modelled. Must be a nonnegative integer. (0, 1, 2, ...) or vector of two nonnegative integers. Represents the number of complete breeding cycles each simulated individual has undergone before the sampling point, where the time between birth and first reproduction is coded as '0', that between first and second reproduction '1', etc. (default 0). Only use in spp. where there is likely to be a reasonable equivalence between breeding stages across a lifespan.}

\item{shape}{(numeric) - value of shape parameter to use with 'vgamma' method. Default 0.5. Must be > 0. Increment towards zero for increasingly heavy-tailed (leptokurtic) dispersal}
}
\value{
returns an object of class \code{KinPairSimulation} containing simulation details and a tibble (tab) of simulation values
}
\description{
Simulate kin dispersal distance pairs with custom species dispersal models.
}
\examples{
custom_dispersal_model <- dispersal_model(a = 10, b = 25, .FS = "b", .HS = "a")
simulate_kindist_custom(nsims = 100, model = custom_dispersal_model, cycle = c(0, 1),
kinship = "FS", lifestage = "b")
}
