% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{spacing}
\alias{spacing}
\alias{spacing<-,KinPairSimulation-method}
\title{Access or alter kin spacing (KinPairSimulation method)}
\usage{
\S4method{spacing}{KinPairSimulation}(x)

\S4method{spacing}{KinPairSimulation}(x) <- value
}
\arguments{
\item{KinPairSimulation}{}
}
\value{



}
\description{
Access or alter kin spacing (KinPairSimulation method)
}
