% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\name{stages}
\alias{stages}
\title{kindisperse - access breeding cycle stages of DispersalModel object.}
\usage{
stages(x)
}
\arguments{
\item{x}{object of class \code{DispersalModel}}
}
\value{
\code{character} ordered vector of custom lifestages contained in the object
}
\description{
kindisperse - access breeding cycle stages of DispersalModel object.
}
