% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairData.R
\name{to_tibble}
\alias{to_tibble}
\title{Extract KinPairData to tibble (generic)}
\usage{
to_tibble(x)
}
\arguments{
\item{x}{object of class \code{KinPairData}}
}
\value{
tibble (class \code{tbl_df})
}
\description{
Extract KinPairData to tibble (generic)
}
