% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ports.R
\name{unmount_appdata}
\alias{unmount_appdata}
\title{Unmount a KinPairData Object (clear slot from appdata environment)}
\usage{
unmount_appdata(nms)
}
\arguments{
\item{nms}{A character vector of names of objects to unmount from the appdata environment}
}
\value{
No return value, called for side effects
}
\description{
Unmount a KinPairData Object (clear slot from appdata environment)
}
\examples{
mount_appdata(KinPairData(), "mydata")

unmount_appdata("mydata")
}
