% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{upper}
\alias{upper}
\alias{upper<-,KinPairSimulation-method}
\title{Access or filter by upper distance (KinPairSimulation method)}
\usage{
\S4method{upper}{KinPairSimulation}(x)

\S4method{upper}{KinPairSimulation}(x) <- value
}
\arguments{
\item{KinPairSimulation}{}
}
\value{



}
\description{
Access or filter by upper distance (KinPairSimulation method)
}
