% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\name{visible_stage}
\alias{visible_stage}
\alias{visible_stage,DispersalModel-method}
\title{Access life stage at which individual is first visible to sampling (i.e. from which breeding cycles are calculated)}
\usage{
visible_stage(x)

\S4method{visible_stage}{DispersalModel}(x)
}
\arguments{
\item{x}{object of class \code{DispersalModel} or \code{}}

\item{DispersalModel}{object of class \code{DispersalModel}}
}
\value{
\code{character} stage in life cycle at which an individual is assumed to be sampled by default rather than its parent
(anchors the breeding cycle system)
}
\description{
Access life stage at which individual is first visible to sampling (i.e. from which breeding cycles are calculated)
}
\section{Methods (by class)}{
\itemize{
\item \code{DispersalModel}: 
}}

