% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R, R/KinPairData.R
\name{breeding_cycle}
\alias{breeding_cycle}
\alias{breeding_cycle,DispersalModel-method}
\alias{breeding_cycle,KinPairData-method}
\title{Access breeding cycle at sampling of \code{\link{DispersalModel}} object.}
\usage{
breeding_cycle(x)

\S4method{breeding_cycle}{DispersalModel}(x)

\S4method{breeding_cycle}{KinPairData}(x)
}
\arguments{
\item{x}{object of class \code{DispersalModel}}

\item{DispersalModel}{object of class \code{DispersalModel}}

\item{KinPairData}{object of class \code{KinPairData}}
}
\value{
\code{integer(s) >= -1} Breeding cycle numbers of modeled dispersed kin. Represents the number of complete
breeding cycles each indivdiual has undergone before the sampling point, where the time between birth and first
reproduction is coded as \code{0}, that between first and second reproduction \code{1}, etc.
}
\description{
Access breeding cycle at sampling of \code{\link{DispersalModel}} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{DispersalModel}: 

\item \code{KinPairData}: 
}}

