% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairData.R
\name{category}
\alias{category}
\alias{category<-}
\title{Access or assign kin category (generic for KinPairData class)}
\usage{
category(x)

category(x) <- value
}
\arguments{
\item{x}{object with relevant method}

\item{value}{new value to assign}
}
\value{
\code{character}. Kinship category of object

returns modified object
}
\description{
Access or assign kin category (generic for KinPairData class)
}
