% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_kinpairs.R
\name{check_valid_lifestage}
\alias{check_valid_lifestage}
\title{Check valid lifestage}
\usage{
check_valid_lifestage(vect)
}
\arguments{
\item{vect}{vector of lifestages}
}
\value{
TRUE if valid. Error otherwise
}
\description{
Checks if vector of lifestages contains all valid entries
}
