% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{csv_to_kinpair}
\alias{csv_to_kinpair}
\title{Reads .csv and converts to KinPairData object}
\usage{
csv_to_kinpair(file, kinship = NULL, lifestage = NULL, ...)
}
\arguments{
\item{file}{The file path to read from}

\item{kinship}{character. kin category to assign or extract from data. one of PO, FS, HS, AV, GG, HAV, GGG, 1C, 1C1, 2C, GAV, HGAV, H1C , H1C1 or H2C}

\item{lifestage}{character. lifestage to assign or extract from data. one of 'unknown', 'immature' or 'ovipositional'.}

\item{...}{additional arguments to pass to read_csv}
}
\value{
returns an object of class \code{KinPairData}
}
\description{
Reads .csv and converts to KinPairData object
}
