% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairData.R
\name{distances}
\alias{distances}
\title{Access distances (generic for KinPairData class)}
\usage{
distances(x)
}
\arguments{
\item{x}{Object of Class KinPairData}
}
\value{
Returns a numeric vector of kin separation distances
}
\description{
Access distances (generic for KinPairData class)
}
