% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{dsigma}
\alias{dsigma}
\alias{juvsigma}
\alias{breedsigma}
\alias{gravsigma}
\alias{ovisigma}
\title{kindisperse - access sigmas of KinPairSimulation objects}
\usage{
dsigma(x)

juvsigma(x)

breedsigma(x)

gravsigma(x)

ovisigma(x)
}
\arguments{
\item{x}{object of class \code{KinPairSimulation}}
}
\value{









}
\description{
kindisperse - access sigmas of KinPairSimulation objects
}
