% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_kindist.R
\name{elongate}
\alias{elongate}
\title{Change the shape (aspect ratio) of a rectangle while preserving area}
\usage{
elongate(dims, aspect = 1)
}
\arguments{
\item{dims}{Original rectangle dimensions - either single number (length of side of square) or length 2 numeric vector (lengths of sides x and y of rectangle)}

\item{aspect}{Aspect ratio of side lengths x & y (i.e. x/y) in the new rectangle}
}
\value{
Returns a numeric vector containing the side lengths c(x, y) of a transformed rectangle with preserved area
}
\description{
This function is used to manipulate the dimensions parameter in other package functions, which control site dimentions.
These geometries can be entered innto functions in a few ways: (a) a single numeric value, which will be interpreted as the length of the side of a square;
(b) a numeric vector of length two, which will be interpreted as the length & width of the sample site; (c) either of the above passed
to this function, which takes the rectangular site dimensions and alters their \code{aspect ratio} (ratio of
length to width) while preserving the underlying area the study site covers.
}
\examples{
elongate(10, 100)
elongate(c(5, 125), 4)
}
