% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\name{fs}
\alias{fs}
\alias{fs,DispersalModel-method}
\title{Access FS phase split point of \code{\link{DispersalModel}} object.}
\usage{
fs(x)

\S4method{fs}{DispersalModel}(x)
}
\arguments{
\item{x}{object of class \code{DispersalModel}}

\item{DispersalModel}{object of class \code{DispersalModel}}
}
\value{
\code{character} FS phase split
}
\description{
Access FS phase split point of \code{\link{DispersalModel}} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{DispersalModel}: 
}}

