% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\name{hs}
\alias{hs}
\alias{hs,DispersalModel-method}
\title{Access HS phase split point of \code{\link{DispersalModel}} object.}
\usage{
hs(x)

\S4method{hs}{DispersalModel}(x)
}
\arguments{
\item{x}{object of class \code{DispersalModel}}

\item{DispersalModel}{object of class \code{DispersalModel}}
}
\value{
\code{character} HS phase split
}
\description{
Access HS phase split point of \code{\link{DispersalModel}} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{DispersalModel}: 
}}

