% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{kerneltype}
\alias{kerneltype}
\title{kindisperse - access kerneltype of KinPairSimulation object}
\usage{
kerneltype(x)
}
\arguments{
\item{x}{object of class \code{KinPairSimulation}}
}
\value{
\code{character} the type of statistical kernel used to run the simulation (Gaussian, Laplace, vgamma)
}
\description{
kindisperse - access kerneltype of KinPairSimulation object
}
