% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindisperse-package.R
\docType{package}
\name{kindisperse-package}
\alias{kindisperse}
\alias{kindisperse-package}
\title{kindisperse: Simulate and Estimate Close-Kin Dispersal Kernels}
\description{
Functions for simulating and estimating kinship-related dispersal. Based
    on the methods described in M. Jasper, T.L. Schmidt., N.W. Ahmad, S.P. Sinkins & A.A. 
    Hoffmann (2019) <doi:10.1111/1755-0998.13043> "A genomic approach to inferring kinship 
    reveals limited intergenerational dispersal in the yellow fever mosquito".
    Assumes an additive variance model of dispersal in two dimensions, compatible with 
    Wright's neighbourhood area. Simple and composite dispersal simulations are supplied, 
    as well as the functions needed to estimate parent-offspring dispersal for simulated or 
    empirical data, and to undertake sampling design for future field studies of dispersal. 
    For ease of use an integrated Shiny app is also included.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moshejasper/kindisperse}
}

}
\author{
\strong{Maintainer}: Moshe-Elijah Jasper \email{moshe.jasper@unimelb.edu.au} (\href{https://orcid.org/0000-0003-4541-3223}{ORCID})

}
\keyword{internal}
