% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindisperse-package.R
\docType{package}
\name{kindisperse-package}
\alias{kindisperse}
\alias{kindisperse-package}
\title{kindisperse: Simulate and Estimate Close-Kin Dispersal Kernels}
\description{
This package implements functions for simulating and estimating kinship-related dispersal, as well as a Shiny app for ease of use.
}
\author{
\strong{Maintainer}: Moshe-Elijah Jasper \email{moshe.jasper@unimelb.edu.au} (\href{https://orcid.org/0000-0003-4541-3223}{ORCID})

}
\keyword{internal}
