% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{kinpair_to_tsv}
\alias{kinpair_to_tsv}
\title{Write KinPairData object in .tsv format (strips extra data)}
\usage{
kinpair_to_tsv(x, file, ...)
}
\arguments{
\item{x}{Object of class KinPairData or KinPairSimulation}

\item{file}{The file path to write to}

\item{...}{Additional arguments to pass to write_tsv}
}
\value{
Invisibly returns the initial object
}
\description{
Write KinPairData object in .tsv format (strips extra data)
}
