% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairData.R
\name{distances,KinPairData-method}
\alias{distances,KinPairData-method}
\alias{kinship,KinPairData-method}
\alias{kinship<-,KinPairData-method}
\alias{lifestage,KinPairData-method}
\alias{lifestage<-,KinPairData-method}
\alias{breeding_cycle,KinPairData-method}
\alias{show,KinPairData-method}
\alias{initialize,KinPairData-method}
\title{Constructor method for \code{KinPairData} objects.}
\usage{
\S4method{distances}{KinPairData}(x)

\S4method{kinship}{KinPairData}(x)

\S4method{kinship}{KinPairData}(x) <- value

\S4method{lifestage}{KinPairData}(x)

\S4method{lifestage}{KinPairData}(x) <- value

\S4method{breeding_cycle}{KinPairData}(x)

\S4method{show}{KinPairData}(object)

\S4method{initialize}{KinPairData}(
  .Object,
  data = NULL,
  kinship = NULL,
  lifestage = NULL,
  cycle = NULL,
  ...
)
}
\arguments{
\item{x}{object of class \code{KinPairData}}

\item{value}{value to assign to slot}

\item{object}{an object of class KinpairData}

\item{.Object}{the KinPairData object to be constructed}

\item{data}{data about kinship to be used to construct object (tibble, data.frame, or numeric vector of distances)}

\item{kinship}{character. Kinship category value for object. - one of PO, FS, HS, AV, HAV, GG, 1C, H1C, GAV, HGAV, 1C1, H1C1, GGG, 2C, and H2C.}

\item{lifestage}{character. Lifestage value for object. - one of 'immature', 'ovipositional' or 'unknown'}

\item{cycle}{non-negative integer or vector of two such integers - Represents
the number of complete breeding cycles each simulated individual has undergone before the sampling point, where the time between
birth and first reproduction is coded as '0', that between first and second reproduction '1', etc. (default 0). If the first individual
was sampled as a juvenile & the second as an adult of equivalent stage, the vector c(0, 1) would be used. In most situations, defualt will be appropriate}

\item{...}{additional argument to pass to downstream functions in future}

\item{KinPairData}{object of class KinPairData}
}
\value{
numeric vector of kin separation distances

\code{character} kinship of \code{KinPairData} object

modified object of class \code{KinPairData}

\code{character} lifestage of \code{KinPairData} object

modified object of class \code{KinPairData}

\code{numeric} number of complete breeding cycles sampled individuals have progressed through

No return value, called for side effects

Returns an object of class \code{KinPairData}
}
\description{
Constructor method for \code{KinPairData} objects.
}
\section{Methods (by generic)}{
\itemize{
\item \code{distances}: access distances

\item \code{kinship}: access kin category

\item \code{kinship<-}: assign kin category

\item \code{lifestage}: access lifestage

\item \code{lifestage<-}: assign lifestage

\item \code{breeding_cycle}: access breeding cycle

\item \code{show}: standard print method

\item \code{initialize}: initialize method
}}

