% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairData.R
\docType{class}
\name{KinPairData-class}
\alias{KinPairData-class}
\alias{KinPairData}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{category}}{character.}

\item{\code{lifestage}}{character.}

\item{\code{tab}}{tbl_df.}
}}

