% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\docType{class}
\name{KinPairSimulation-class}
\alias{KinPairSimulation-class}
\alias{KinPairSimulation}
\title{KinPairSimulation}
\value{
Returns an object of class 'KinPairSimulation'
}
\description{
KinPairSimulation
}
\section{Slots}{

\describe{
\item{\code{category}}{character - one of PO, FS, HS, AV, HAV, GG, 1C, H1C, GAV, HGAV, 1C1, H1C1, GGG, 2C, and H2C.}

\item{\code{simtype}}{character.}

\item{\code{kerneltype}}{character. - 'Gaussian' or 'Laplace'}

\item{\code{sigma}}{numeric.       - overall value of dispersal sigma (for simple kernel)}

\item{\code{juvsigma}}{numeric.    - value of pre-breeding dispersal sigma (for composite kernel)}

\item{\code{breedsigma}}{numeric.  - value of breeding dispersal sigma (for composite kernel)}

\item{\code{gravsigma}}{numeric.   - value of post-breeding dispersal sigma (for composite kernel)}

\item{\code{ovisigma}}{numeric.    - value of oviposition dispersal sigma (for composite kernel)}

\item{\code{dims}}{numeric.        - dimensions of sampling area (assumes 1 side of square)}

\item{\code{lifestage}}{character. - lifestage at sampling - either 'larva' or 'oviposition'}

\item{\code{call}}{call.           - call to create initial simulation}

\item{\code{tab}}{tbl_df.          - tibble of simulation values}

\item{\code{filtertype}}{character. - whether the initial sim has been further filtered}

\item{\code{upper}}{numeric.       - FILTER: upper threshold used}

\item{\code{lower}}{numeric.       - FILTER: lower threshold used}

\item{\code{spacing}}{numeric.     - FILTER: spacing used}

\item{\code{samplenum}}{numeric.   - FILTER: sample number used}

\item{\code{sampledims}}{numeric.  - FILTER: dimensions used}
}}

