% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairData.R
\name{lifestage}
\alias{lifestage}
\alias{lifestage<-}
\title{Access or assign lifestage (generic for KinPairData class)}
\usage{
lifestage(x)

lifestage(x) <- value
}
\arguments{
\item{x}{object with relevant method}

\item{value}{new value to assign}
}
\value{
\code{character} life stage of object

returns modified object with altered lifestage
}
\description{
Access or assign lifestage (generic for KinPairData class)
}
