% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairData.R
\name{lifestage}
\alias{lifestage}
\alias{lifestage<-,KinPairData-method}
\title{Title}
\usage{
lifestage(x)

\S4method{lifestage}{KinPairData}(x)

\S4method{lifestage}{KinPairData}(x) <- value
}
\arguments{
\item{x}{}

\item{KinPairData}{}
}
\value{





}
\description{
Title

Access or assign KinPairData lifestage
}
