% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{read_kindata}
\alias{read_kindata}
\title{Reads .kindata filetype back to KinPairData or KinPairSimulation object.}
\usage{
read_kindata(file)
}
\arguments{
\item{file}{Character giving path reference to file with extension .kinpair}
}
\value{
Returns either \code{KinPairData} or \code{KinPairSimulation} object.
}
\description{
Reads .kindata filetype back to KinPairData or KinPairSimulation object.
}
