% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ports.R
\name{reset_appdata}
\alias{reset_appdata}
\title{Reset kindisperse appdata}
\usage{
reset_appdata()
}
\value{
No return value, called for side effects
}
\description{
Reset kindisperse appdata
}
\examples{
reset_appdata()
}
