% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ports.R
\name{reset_tempdata}
\alias{reset_tempdata}
\title{Reset app tempdata (internal mem)}
\usage{
reset_tempdata()
}
\value{
No return value, called for side effects
}
\description{
Reset app tempdata (internal mem)
}
\examples{
reset_tempdata()
}
