% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ports.R
\name{retrieve_tempdata}
\alias{retrieve_tempdata}
\title{Retrieve all tempdata (internal mem) from app (as list)}
\usage{
retrieve_tempdata()
}
\value{
A list of all KinPairData objects in kindisperse app's tempdata
}
\description{
Retrieve all tempdata (internal mem) from app (as list)
}
\examples{
retrieve_tempdata()
}
