% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R, R/KinPairData.R
\name{sampling_stage}
\alias{sampling_stage}
\alias{sampling_stage<-}
\alias{sampling_stage,DispersalModel-method}
\alias{sampling_stage<-,DispersalModel-method}
\alias{sampling_stage,KinPairData-method}
\title{Access sampling stage of \code{\link{DispersalModel}} or \code{\link{KinPairSimulation}} object.}
\usage{
sampling_stage(x)

sampling_stage(x) <- value

\S4method{sampling_stage}{DispersalModel}(x)

\S4method{sampling_stage}{DispersalModel}(x) <- value

\S4method{sampling_stage}{KinPairData}(x)
}
\arguments{
\item{x}{object of class \code{DispersalModel}}

\item{value}{\code{character} new sampling stage to assign model}

\item{DispersalModel}{object of class \code{DispersalModel}}

\item{KinPairData}{object of class \code{KinPairData}}
}
\value{
\code{character} sampling stage

returns a modified object of class \code{DispersalModel}
}
\description{
Access sampling stage of \code{\link{DispersalModel}} or \code{\link{KinPairSimulation}} object.
}
\section{Methods (by class)}{
\itemize{
\item \code{DispersalModel}: 

\item \code{KinPairData}: 
}}

