% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DispersalModel.R
\name{sampling_stage}
\alias{sampling_stage}
\title{kindisperse - access sampling stage of DispersalModel object.}
\usage{
sampling_stage(x)
}
\arguments{
\item{x}{object of class \code{DispersalModel}}
}
\value{
\code{character} sampling stage
}
\description{
kindisperse - access sampling stage of DispersalModel object.
}
