% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{simdims}
\alias{simdims}
\title{Access simulation dimensions of KinPairSimulation object}
\usage{
simdims(x)
}
\arguments{
\item{x}{object of class \code{KinPairSimulation}}
}
\value{
\code{numeric vector} dimensions of simulated object
}
\description{
Access simulation dimensions of KinPairSimulation object
}
