% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{simdims}
\alias{simdims}
\title{Access simulation dimensions (KinPairSimulation method)}
\usage{
\S4method{simdims}{KinPairSimulation}(x)
}
\arguments{
\item{KinPairSimulation}{}
}
\value{

}
\description{
Access simulation dimensions (KinPairSimulation method)
}
