% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{simtype}
\alias{simtype}
\title{kindisperse - access simtype of KinPairSimulation object}
\usage{
simtype(x)
}
\arguments{
\item{x}{object of class \code{KinPairSimulation}}
}
\value{
\code{character} the kind of simulation stored in the object (simple or composite)
}
\description{
kindisperse - access simtype of KinPairSimulation object
}
