% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{upper}
\alias{upper}
\alias{upper<-}
\alias{lower}
\alias{lower<-}
\alias{spacing}
\alias{spacing<-}
\alias{samplenum}
\alias{samplenum<-}
\alias{sampledims}
\alias{sampledims<-}
\title{Access & filter by filter parameters of KinPairSimulation Object}
\usage{
upper(x)

upper(x) <- value

lower(x)

lower(x) <- value

spacing(x)

spacing(x) <- value

samplenum(x)

samplenum(x) <- value

sampledims(x)

sampledims(x) <- value
}
\arguments{
\item{x}{object of class \code{KinPairSimulation}}

\item{value}{new value to assign}
}
\value{
\code{numeric} upper value of sampled object

returns a modified object of the relevant class

\code{numeric} lower value of sampled object

returns a modified object of the relevant class

\code{numeric} trap spacing of sampled object

returns a modified object of the relevant class

\code{numeric} number of kin dyads in sampled object

returns a modified object of the relevant class

\code{numeric vector} dimensions of sampled object

returns a modified object of the relevant class
}
\description{
Access & filter by filter parameters of KinPairSimulation Object
}
