% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KinPairSimulation.R
\name{upper}
\alias{upper}
\alias{upper<-}
\alias{lower}
\alias{lower<-}
\alias{spacing}
\alias{spacing<-}
\alias{samplenum}
\alias{samplenum<-}
\alias{sampledims}
\alias{sampledims<-}
\title{Access & filter by filter parameters of KinPairSimulation Object}
\usage{
upper(x)

upper(x) <- value

lower(x)

lower(x) <- value

spacing(x)

spacing(x) <- value

samplenum(x)

samplenum(x) <- value

sampledims(x)

sampledims(x) <- value
}
\arguments{
\item{x}{object of class \code{KinPairSimulation}}

\item{value}{new value to assign}
}
\value{



















}
\description{
Access & filter by filter parameters of KinPairSimulation Object
}
