% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_functions.R
\name{write_kindata}
\alias{write_kindata}
\title{Write KinPairData or KinPairSimulation object in .kindata format (no information lost)}
\usage{
write_kindata(x, file)
}
\arguments{
\item{x}{Object of class KinPairData or KinPairSimulation}

\item{file}{The file path to write to. If is doesn't end it '.kindata', this will be added.}
}
\value{
Invisibly returns the initial object
}
\description{
Write KinPairData or KinPairSimulation object in .kindata format (no information lost)
}
